+++ backend/src/application/dto/stats_dto.rs
@@ -1,4 +1,5 @@
 use chrono::{DateTime, Utc};
+use rust_decimal::Decimal;
 use serde::Serialize;
 
 #[derive(Debug, Clone, Serialize)]
@@ -39,7 +40,18 @@ pub struct SyndicDashboardStats {
     pub total_units: i64,
     pub total_owners: i64,
     pub pending_expenses_count: i64,
-    pub pending_expenses_amount: f64,
+    /// Total des dépenses en attente, en EUR.
+    ///
+    /// `Decimal` : la colonne `expenses.amount` est en `NUMERIC(12,2)` depuis
+    /// la migration 20260502000000. La lire en `f64` était une dégradation
+    /// gratuite d'une valeur déjà exacte (même défaut que celui trouvé dans
+    /// `find_overdue_expenses_without_reminders` sous #661).
+    ///
+    /// `serde::float` conserve la représentation JSON numérique attendue par
+    /// `OwnerDashboard.svelte` / `SyndicDashboard.svelte`, qui typent ce champ
+    /// `number` — aucun drift de contrat.
+    #[serde(with = "rust_decimal::serde::float")]
+    pub pending_expenses_amount: Decimal,
     pub next_meeting: Option<NextMeetingInfo>,
 }
